IDs as u16.

+++ src/ids.rs
@@ -1,7 +1,7 @@
-pub const MAIN_ICON:   i32 = 101;
+pub const MAIN_ICON:   u16 = 101;
 
-pub const DIALOG_MAIN: i32 = 1000;
-pub const BTN_PATCH:   i32 = 1001;
-pub const LBL_PATH:    i32 = 1002;
-pub const TXT_PATH:    i32 = 1003;
-pub const BTN_CHOOSE:  i32 = 1004;
+pub const DIALOG_MAIN: u16 = 1000;
+pub const BTN_PATCH:   u16 = 1001;
+pub const LBL_PATH:    u16 = 1002;
+pub const TXT_PATH:    u16 = 1003;
+pub const BTN_CHOOSE:  u16 = 1004;
